feat(selection_mode/line): left = parent line
This also entails no more generic Parent Line movements.
