+++ docs/docs/components/file-explorer.md
@@ -41,15 +41,15 @@ Other keybindings can be found at [contextual keybindings](../normal-mode/space-
 
 Because the File Explorer is just a YAML file, the following actions are free[^1]:
 
-| Action                                                                        | How?                                         |
-| ----------------------------------------------------------------------------- | -------------------------------------------- |
-| Go to parent folder                                                           | Use [Parent Line][1]                         |
-| Go to first/last file in current folder                                       | Use [First/Last][2] with [Syntax Node][3]    |
+| Action                                                                       | How?                                         |
+| ---------------------------------------------------------------------------- | -------------------------------------------- |
+| Go to parent folder                                                          | Use [`e h`][1]                               |
+| Go to first/last file in current folder                                      | Use [First/Last][2] with [Syntax Node][3]    |
 | Go to next/previous file/folder at current level, skipping expanded children | Use [Previous/Next][4] with [Syntax Node][3] |
 
 [^1]: Free as in no extra implementations required
 
-[1]: ../normal-mode/core-movements.mdx#parent-line
-[2]: ../normal-mode/core-movements.mdx#firstlast
+[1]: ../normal-mode/selection-modes/regex-based.md#line
+[2]: ../normal-mode/core-movements.md#firstlast
 [3]: ../normal-mode/selection-modes/syntax-node-based.md#syntax-node
-[4]: ../normal-mode/core-movements.mdx#leftright
+[4]: ../normal-mode/core-movements.md#leftright
