+++ docs/docs/normal-mode/actions/index.md
@@ -141,7 +141,7 @@ Reason: The `esc enter` combo is sweet.
 
 Upon saving, formatting will be applied if possible.
 
-After formatting, the [Current](../core-movements.mdx#current) movement will be executed, to reduce disorientation caused by the misplaced selection due to content changes.
+After formatting, the [Current](../core-movements.md#current) movement will be executed, to reduce disorientation caused by the misplaced selection due to content changes.
 
 ## Undo/Redo
