+++ docs/docs/normal-mode/selection-modes/index.md
@@ -4,12 +4,12 @@ sidebar_position: 1
 
 # Selection Modes
 
-Selection modes [^1] dictates how [core movements](../core-movements.mdx) works.
+Selection modes [^1] dictates how [core movements](../core-movements.md) works.
 
 There are roughly 3 categories of selection modes (not clear-cut):
 
 1. [Syntax tree-based](./syntax-node-based.md)
-2. [Regex-based](./regex-based.mdx)
+2. [Regex-based](./regex-based.md)
 3. [Local/Global](./local-global/index.md)
 
 [^1]: For Vim users, selection mode means text objects.
