+++ docs/docs/normal-mode/selection-modes/regex-based.md
@@ -10,11 +10,20 @@ import {TutorialFallback} from '@site/src/components/TutorialFallback';
 
 Keybinding: `e`
 
-In this selection mode `h`/`l` behaves exactly like `j`/`k`, and the selection
-is trimmed, which means that the leading and trailing spaces of each line are
-not selected.
+In this selection mode, the selection is trimmed, which means that the leading
+and trailing spaces of each line are not selected.
 
 This is usually used in conjunction with `i`/`a` to immediately enter insert mode at the first/last non-whitespace symbol of the current line.
+
+`h` moves the nearest parent line.
+
+Parent lines are highlighted lines that represent the parent nodes of the current selection.
+
+This is useful for example when you are within the body of a function and you want to jump to the function name.
+
+This is also practical in the [File Explorer](../../components/file-explorer.md) because the file explorer is rendered using YAML, so going to Parent Line means going to the parent folder!
+
+<TutorialFallback filename="line"/>
 
 ## Full Line
 
@@ -51,6 +60,6 @@ Keybindings:
 - `z`: Collapse selection (start)
 - `$`: Collapse selection (end)
 
-In this selection mode, the movements behave like the usual editor, where [Left/Right](./../core-movements.mdx#leftright) means left/right, and so on.
+In this selection mode, the movements behave like the usual editor, where [Left/Right](./../core-movements.md#leftright) means left/right, and so on.
 
-[First/Last](./../core-movements.mdx#firstlast) means the first/last column of the current line.
+[First/Last](./../core-movements.md#firstlast) means the first/last column of the current line.
