+++ src/selection_mode/line_trimmed.rs
@@ -1,4 +1,6 @@
-use super::SelectionMode;
+use crate::components::editor::IfCurrentNotFound;
+
+use super::{SelectionMode, SelectionModeParams};
 
 pub(crate) struct LineTrimmed;
 
@@ -37,6 +39,39 @@ impl SelectionMode for LineTrimmed {
                 }),
         ))
     }
+
+    fn left(
+        &self,
+        params: super::SelectionModeParams,
+    ) -> anyhow::Result<Option<crate::selection::Selection>> {
+        let SelectionModeParams {
+            buffer,
+            current_selection,
+            cursor_direction,
+        } = params;
+        let current_line = buffer.char_to_line(current_selection.extended_range().start)?;
+        Ok(buffer
+            .get_parent_lines(current_line)?
+            .into_iter()
+            .filter(|line| line.line < current_line)
+            .next_back()
+            .map(|line| {
+                let byte_range = buffer.line_to_byte_range(line.line)?;
+                let start = trim_leading_spaces(byte_range.range.start, &line.content);
+                let char_index_range =
+                    buffer.byte_range_to_char_index_range(&(start..start + 1))?;
+                self.current(
+                    SelectionModeParams {
+                        buffer,
+                        cursor_direction,
+                        current_selection: &current_selection.clone().set_range(char_index_range),
+                    },
+                    IfCurrentNotFound::LookForward,
+                )
+            })
+            .transpose()?
+            .flatten())
+    }
 }
 
 pub(crate) fn trim_leading_spaces(byte_start: usize, line: &str) -> usize {
@@ -54,7 +89,7 @@ pub(crate) fn trim_leading_spaces(byte_start: usize, line: &str) -> usize {
 
 #[cfg(test)]
 mod test_line {
-    use crate::{buffer::Buffer, selection::Selection};
+    use crate::{buffer::Buffer, components::editor::Direction, selection::Selection};
 
     use super::*;
 
@@ -82,5 +117,42 @@ mod test_line {
     fn single_line_without_trailing_newline_character() {
         let buffer = Buffer::new(None, "a");
         LineTrimmed.assert_all_selections(&buffer, Selection::default(), &[(0..1, "a")]);
+    }
+
+    #[test]
+    fn to_parent_line() {
+        let buffer = Buffer::new(
+            Some(tree_sitter_rust::language()),
+            "
+fn f() {
+    fn g() {
+        let a = 1;
+        let b = 2;
+        let c = 3;
+        let d = 4;
+    }
+
+}"
+            .trim(),
+        );
+
+        let test = |selected_line: usize, expected: &str| {
+            let start = buffer.line_to_char(selected_line).unwrap();
+            let result = LineTrimmed
+                .left(SelectionModeParams {
+                    buffer: &buffer,
+                    current_selection: &Selection::new((start..start + 1).into()),
+                    cursor_direction: &Direction::default(),
+                })
+                .unwrap()
+                .unwrap();
+
+            let actual = buffer.slice(&result.extended_range()).unwrap();
+            assert_eq!(actual, expected);
+        };
+
+        test(4, "fn g() {");
+
+        test(1, "fn f() {");
     }
 }
